clean up tablet mode.  Finish mobile view.  Version up
