+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "frontend-apps",
-  "version": "0.4.5",
+  "version": "0.4.6",
   "license": "Apache-2.0",
   "private": false,
   "dependencies": {
