+++ src/Components/Menu/MenuMobile.js
@@ -1,48 +1,58 @@
-import React from 'react';
+import React, { useState } from 'react';
 import styled, { keyframes } from 'styled-components';
 import PropTypes from 'prop-types';
-import MenuItem from './MenuItem';
+import { breakpoints } from 'consts';
+import { default as BaseSprite } from '../Sprite';
+import { default as BaseMenuItem } from './MenuItem';
+import { default as BaseDropdownList } from '../DropdownList';
 import MenuLine from './MenuLine';
 import logo from './logo.svg';
 
 const MenuContainer = styled.div`
-  min-width: 400px;
-  width: 400px;
-  position: fixed;
-  overflow: hidden;
-  top: 0;
-  left: 0;
-  height: 100%;
-  min-height: 100vh;
+  width: 100%;
   text-transform: uppercase;
   font-size: 1.2rem;
   letter-spacing: 0.24rem;
-  box-shadow: 0 4px 50px 20px ${({ theme }) => theme.BLACK95 };
-  z-index: 20;
 `;
 const MenuContent = styled.div`
-  width: 415px;
-  padding: 50px;
-  height: 100%;
-  background: ${({ theme }) => theme.MENU_BG };
-  padding-bottom: 100px;
-  overflow-x: hidden;
-  overflow-y: scroll;
+  width: 100%;
+  padding: 30px 30px 20px 30px;
+  position: fixed;
+  top: 0;
+  left: 0;
+  background: ${({ theme }) => theme.BLACK };
+  box-shadow: 0 5px 20px 0px ${({ theme }) => theme.BLACK95 };
+  z-index: 20;
+  @media ${breakpoints.down('sm')} {
+    padding: 30px 15px 20px 15px;
+  }
 `;
 const MenuHeader = styled.div`
-  margin-bottom: 20px;
 `;
 const MenuLogo = styled.div`
   display: flex;
   align-items: center;
   position: relative;
   cursor: pointer;
-  margin-bottom: 40px;
+  margin-bottom: 30px;
 `;
 const LogoTitle = styled.h1`
   line-height: 1.4rem;
 `;
-const MenuItems = styled.div``;
+const MenuItems = styled.div`
+  background: ${({ theme }) => theme.GREY_DARKEST };
+  position: fixed;
+  top: 0;
+  left: 0;
+  height: 100vh;
+  width: 100vw;
+  overflow-y: scroll;
+  padding: 30px 15px;
+  z-index: 50;
+`;
+const MenuItem = styled(BaseMenuItem)`
+  width: 300px;
+`;
 const MenuSection = styled.div`
   position: relative;
   ${({ indent }) => indent && `margin-left: ${indent}px;` }
@@ -85,63 +95,121 @@ const Logo = styled.img`
   width: 62px;
   margin-right: 20px;
 `;
+const Sprite = styled(BaseSprite)`
+  position: absolute;
+  top: 30px;
+  right: 15px;
+`;
+const DropdownList = styled(BaseDropdownList)`
+  @media ${breakpoints.down('sm')} {
+    width: 100%;
+  }
+`;
 
-const MenuMobile = ({ className }) => (
-  <MenuContainer className={className}>
-    <MenuContent>
-      <MenuHeader>
-        <MenuLogo onClick={() => { window.scrollTo({top: 0, behavior: 'smooth'}) }}>
-          <Logo src={logo} />
-          <LogoTitle>Provenance Applications Mobile</LogoTitle>
-          {process.env.REACT_APP_ENV !== 'production' && <TestnetMessage>TESTNET</TestnetMessage>}
-          <Version>v{process.env.REACT_APP_VERSION}</Version>
-        </MenuLogo>
+const MenuTablet = ({ className }) => {
+  const [menuOpen, setMenuOpen] = useState(false);
+  const [dropdownValue, setDropdownValue] = useState('wallet');
+  
+  const toggleMenu = () => {
+    const bodyElement = document.getElementsByTagName('body')[0];
+    // If the menu is open, add a max-height and overflow hidden to the body (no scrolling past open menu)
+    if (menuOpen) { // Close Menu
+      bodyElement.style.overflow = '';
+      bodyElement.style.maxHeight = '';
+      setMenuOpen(false);
+    } else { // Open Menu
+      bodyElement.style.overflow = 'hidden';
+      bodyElement.style.maxHeight = '100%';
+      setMenuOpen(true);
+    }
+  };
+
+  const scrollTo = (section) => {
+    const targetElement = document.getElementById(section);
+    if (targetElement) {
+      // We need to scroll with an offset since the header floats and blocks content
+      const offset = 200;
+      const bodyRect = document.body.getBoundingClientRect().top;
+      const elementRect = targetElement.getBoundingClientRect().top;
+      const elementPosition = elementRect - bodyRect;
+      const scrollPosition = elementPosition - offset;
+      window.scrollTo({ behavior: 'smooth', top: scrollPosition });
+    }
+  };
+
+  const menuItemClick = (section) => {
+    toggleMenu();
+    setDropdownValue(section);
+    scrollTo(section);
+  };
+
+  const changeDropdown = ({ target }) => {
+    setDropdownValue(target.value);
+    scrollTo(target.value);
+  };
+
+  return (
+    <MenuContainer className={className}>
+      <MenuContent>
+        <MenuHeader>
+          <MenuLogo onClick={() => { window.scrollTo({top: 0, behavior: 'smooth'}) }}>
+            <Logo src={logo} />
+            <LogoTitle>Provenance Applications</LogoTitle>
+            {process.env.REACT_APP_ENV !== 'production' && <TestnetMessage>TESTNET</TestnetMessage>}
+            <Version>v{process.env.REACT_APP_VERSION}</Version>
+          </MenuLogo>
+          <Sprite icon="MENU" size="2rem" color="WHITE" onClick={toggleMenu} />
+          <DropdownList value={dropdownValue} onChange={changeDropdown} />
+        </MenuHeader>
+      </MenuContent>
+      {menuOpen && (
         <MenuItems>
+          <Sprite icon="CLOSE" onClick={toggleMenu} color="WHITE" size="1.5rem" />
           <MenuSection>
             <MenuLine length="446" top="20" left="3" direction="down" />
             <SectionTitle>Wallet</SectionTitle>
-            <MenuItem title="Create a Wallet" section="wallet" indent="30" />
+            <MenuItem title="Create a Wallet" onClick={() => menuItemClick("wallet")} indent="30" />
           </MenuSection>
           <MenuSection indent="30">
             <MenuLine length="14" top="3" left="-24" direction="left" />
             <SectionTitle>NFT</SectionTitle>
-            <MenuItem title="Create an NFT" section="nft" />
-            <MenuItem title="Tokenize an NFT" section="nft" />
+            <MenuItem title="Create an NFT" onClick={() => menuItemClick('nft')} />
+            <MenuItem title="Tokenize an NFT" onClick={() => menuItemClick('nft')} />
           </MenuSection>
           <MenuSection indent="30">
             <MenuLine length="14" top="3" left="-24" direction="left" />
             <SectionTitle>Hash</SectionTitle>
-            <MenuItem title="Purchase Hash" section="hash" />
-            <MenuItem title="Delegate Hash" section="hash" />
-            <MenuItem title="Transfer Hash" section="hash" />
+            <MenuItem title="Purchase Hash" onClick={() => menuItemClick('hash')} />
+            <MenuItem title="Delegate Hash" onClick={() => menuItemClick('hash')} />
+            <MenuItem title="Transfer Hash" onClick={() => menuItemClick('hash')} />
           </MenuSection>
           <MenuSection indent="30">
             <MenuLine length="153" top="3" left="-10" />
             <MenuLine length="14" top="3" left="-24" direction="left"  />
             <SectionTitle>Passport</SectionTitle>
-            <MenuItem title="Add a Passport" section="passport" />
-            <MenuItem title="Upgrade Passport" section="passport" />
+            <MenuItem title="Add a Passport" onClick={() => menuItemClick('passport')} />
+            <MenuItem title="Upgrade Passport" onClick={() => menuItemClick('passport')} />
           </MenuSection>
           <MenuSection indent="42">
             <MenuLine length="10" top="3" left="-20" direction="left" />
             <SectionTitle>Exchange</SectionTitle>
-            <MenuItem title="Subscribe to a Fund" section="exchange" />
-            <MenuItem title="Buy Digital Currency" section="exchange" />
-            <MenuItem title="Trade on ATS" section="exchange" />
-            <MenuItem title="Access BTC Bridge" section="exchange" />
-            <MenuItem title="Access ETH Bridge" section="exchange" />
+            <MenuItem title="Subscribe to a Fund" onClick={() => menuItemClick('exchange')} />
+            <MenuItem title="Buy Digital Currency" onClick={() => menuItemClick('exchange')} />
+            <MenuItem title="Trade on ATS" onClick={() => menuItemClick('exchange')} />
+            <MenuItem title="Access BTC Bridge" onClick={() => menuItemClick('exchange')} />
+            <MenuItem title="Access ETH Bridge" onClick={() => menuItemClick('exchange')} />
           </MenuSection>
         </MenuItems>
-      </MenuHeader>
-    </MenuContent>
-  </MenuContainer>
-);
+      )}
+    </MenuContainer>
+  );
+};
 
-MenuMobile.propTypes = {
+MenuTablet.propTypes = {
   className: PropTypes.string,
 };
-MenuMobile.defaultProps = {
+MenuTablet.defaultProps = {
   className: '',
 };
 
-export default MenuMobile;
+export default MenuTablet;
