+++ src/Components/Menu/index.js
@@ -1,3 +1,1 @@
 export { default as Menu } from './Menu';
-export { default as MenuTablet } from './MenuTablet';
-export { default as MenuMobile } from './MenuMobile';
